Stable Diffusion as AMI
Motivation
---
We want to run stable-diffusion in AWS on demand as an EC2 instance.

Modifications
---
- Updated startup script template that runs as part of the linux bootup
- Define the EC2 instance in terraform
- Declare variables used by terraform

Note(s)
---
These scripts won't be useful out of the box and will require adjusting variables for your own use.

In particular, the stable-diffusion AMI defined in the variables is built in a separate project and deployed to AWS separately. However, this should provide a guide for how you might go about doing this for yourself.